+++ testcase/bvd_add/case_bvd_add_04/validate_result.sh
@@ -2,8 +2,11 @@
 
 scriptdir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
-rm $scriptdir/../../../bin/DB/*
-rmdir $scriptdir/../../../bin/DB/
+if [ -e $scriptdir/../../../bin/DB/bvdb ]
+then
+	rm $scriptdir/../../../bin/DB/*
+	rmdir $scriptdir/../../../bin/DB/
+fi
 
 $scriptdir/../../../bin/bvd-add.pl $scriptdir/case_bvd_add_04_1.vcf
 $scriptdir/../../../bin/bvd-add.pl $scriptdir/case_bvd_add_04_2.vcf -T colon_cancer,lung_cancer
